Split up type defintions
