+++ server/graphql/backend.js
@@ -1,0 +1,1 @@
+backend.js
