+++ server/graphql/types.js
@@ -1,0 +1,7 @@
+import {Region} from './types/basic'
+import {Collection} from './types/collections'
+
+export default {
+	Region,
+	Collection
+}
