+++ server/graphql/types/basic.js
@@ -1,0 +1,20 @@
+import {GraphQLEnumType} from 'graphql';
+
+const Region = new GraphQLEnumType({
+	name: "Region",
+	description: "Region with specific content",
+	values: {
+		UK: {
+			value: 'uk',
+			description: "United Kingdom"
+		},
+		US: {
+			value: 'us',
+			description: "United States of America"
+		}
+	}
+});
+
+export default {
+	Region
+}
